subgraph: refactor test
